+++ src/App.tsx
@@ -8,13 +8,13 @@ import Positions from '@/pages/positions/Positions'
 import Library from '@/pages/Library'
 import Requests from '@/pages/Requests'
 
-import JobDetails from '@/pages/applicants/JobDetails'
-import LeadDeveloper from '@/pages/applicants/lead-developer'
+import JobDetails from '@/pages/applicants/Job'
+import LeadDeveloper from '@/pages/applicants/JobDetails'
 import LeadDeveloperWeekly from '@/pages/applicants/lead-developer-weekly-monthly'
 import ApplicantInformationTab from '@/pages/applicants/ApplicantsInformationTab'
-import LeadDeveloperRS from '@/pages/applicants/LeadDeveloper_ResumeScreening'
-import LeadDeveloperPI from '@/pages/applicants/LeadDeveloper_PhoneInterview'
-import LeadDeveloperSL from '@/pages/applicants/LeadDeveloper_Shortlisted'
+import LeadDeveloperRS from '@/pages/applicants/ResumeScreening'
+import LeadDeveloperPI from '@/pages/applicants/PhoneInterview'
+import LeadDeveloperSL from '@/pages/applicants/Shortlisted'
 
 import InterviewEvaluationForm from '@/pages/InterviewEvaluationFormPage'
 
@@ -23,11 +23,12 @@ import PositionsOnHold from '@/pages/positions/positions-onhold'
 import PositionsPublished from '@/pages/positions/positions-published'
 import PositionsClosed from '@/pages/positions/positions-closed'
 import PositionsArchive from '@/pages/positions/positions-archive'
-import LeadDeveloperII from '@/pages/LeadDeveloper_InitialInterview'
-import LeadDeveloperAS from '@/pages/LeadDeveloper_Assessment'
-import LeadDeveloperFI from '@/pages/LeadDeveloper_FinalInterview'
-import LeadDeveloperFJO from '@/pages/LeadDeveloper_ForJobOffer'
+import LeadDeveloperII from '@/pages/InitialInterview'
+import LeadDeveloperAS from '@/pages/Assessment'
+import LeadDeveloperFI from '@/pages/FinalInterview'
+import LeadDeveloperFJO from '@/pages/ForJobOffer'
 import OfferAndFinalization from '@/pages/OfferAndFinalization'
+import JobStageTemplate from './pages/applicants/JobStateTemplate'
 
 
 export default function App() {
@@ -39,17 +40,17 @@ export default function App() {
 
       {/* Main Tabs */}
       <Route path="/dashboard/*" element={<Page />} />
-      <Route path="/applicants/*" element={<Applicants />} />
+      <Route path="/applicants/" element={<Applicants />} />
       <Route path="/positions/*" element={<Positions />} />
       <Route path="/library/*" element={<Library />} />
       <Route path="/requests/*" element={<Requests />} />
 
       {/* Applicants */}
-      <Route path="/applicants/jobdetails" element={<JobDetails />} />
-      <Route path="/applicants/jobdetails/leaddeveloper" element={<LeadDeveloper />} />
+      <Route path="/applicants/job" element={<JobDetails />} />
+      <Route path="/applicants/job/:jobtitle" element={<LeadDeveloper />} />
       <Route path="/applicants/jobdetails/leaddeveloper/leaddeveloperweekly" element={<LeadDeveloperWeekly />} />
       <Route path="/ApplicantInformation/" element={<ApplicantInformationTab />} />
-      <Route path="/applicants/jobdetails/leaddeveloper/LeadDeveloperRS" element={<LeadDeveloperRS />} />
+      <Route path="/applicants/jobdetails/leaddeveloper/resumescreening" element={<LeadDeveloperRS />} />
       <Route path="/applicants/jobdetails/leaddeveloper/LeadDeveloperPI" element={<LeadDeveloperPI/>} />
       <Route path="/applicants/jobdetails/leaddeveloper/LeadDeveloperSL" element={<LeadDeveloperSL/>} />
 
@@ -67,6 +68,7 @@ export default function App() {
       <Route path="/applicants/jobdetails/leaddeveloper/LeadDeveloperFI" element={<LeadDeveloperFI/>} />
       <Route path="/applicants/jobdetails/leaddeveloper/LeadDeveloperFJO" element={<LeadDeveloperFJO/>} />  
       <Route  path="/applicants/OfferAndFinalization" element={<OfferAndFinalization />} />
+      <Route path="/applicants/job/:jobtitle/:jobstage" element={<JobStageTemplate />} />
       </Routes>
     </BrowserRouter>
   );
