+++ src/pages/InitialInterview.tsx
@@ -14,6 +14,8 @@ import { Label } from "@/components/ui/label"
 import { Textarea } from "@/components/ui/textarea"
 import { Card, CardContent } from "@/components/ui/card"
 import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group"
+import { useParams } from "react-router-dom"
+
 
 // Sample applicant data
 const applicants = [
@@ -133,6 +135,8 @@ function Sidebar() {
         return "bg-gray-100 text-gray-800 border-gray-200"
     }
   }
+
+
 
   return (
     <div className="w-full xl:w-[320px] bg-white xl:border-l border-gray-200 p-4 pt-4 space-y-4 xl:sticky xl:top-0 xl:h-screen xl:overflow-y-auto">
@@ -419,6 +423,26 @@ export default function JobManagement() {
     applicant.name.toLowerCase().includes(searchTerm.toLowerCase()),
   )
 
+  
+      const formatJobTitle = (slug?: string) => {
+  const titleMap: Record<string, string> = {
+    leaddeveloper: "Lead Developer",
+    projectmanager: "Project Manager",
+    socialcontentmanager: "Social Content Manager",
+    senioruiuxdesigner: "Senior UI/UX Designer",
+    customersupport: "Customer Support",
+    qaengineer: "QA Engineer",
+    humanresourcescoordinator: "Human Resources Coordinator",
+    operationsmanager: "Operations Manager",
+    socialmediamanager: "Social Media Manager",
+    marketingspecialist: "Marketing Specialist",
+    seniorsoftwareengineer: "Senior Software Engineer",
+  }
+  return slug ? titleMap[slug.toLowerCase()] || slug.replace(/([a-z])([A-Z])/g, "$1 $2") : "Unknown Job"
+}
+const { jobtitle } = useParams<{ jobtitle: string }>()
+const resolvedJobTitle = formatJobTitle(jobtitle)
+
   return (
     <>
       <Navbar />
@@ -442,9 +466,8 @@ export default function JobManagement() {
                   </Button>
 
                   <div className="flex flex-col items-center sm:flex-row sm:items-center sm:gap-3">
-                    <h1 className="text-xl font-bold text-gray-900 text-center sm:text-left">
-                      Senior Software Engineer
-                    </h1>
+                    <h1 className="text-2xl font-bold text-gray-900 text-center sm:text-left">{resolvedJobTitle}</h1>
+
                     <Select value={jobStatus} onValueChange={setJobStatus}>
                       <SelectTrigger
                         className={`mt-10 sm:mt-0 w-auto min-w-[80px] text-xs font-medium border rounded px-2 py-0.5 ${statusStyles[jobStatus as keyof typeof statusStyles]}`}
@@ -479,50 +502,41 @@ export default function JobManagement() {
                 {/* Filter and Search */}
                 <div className="flex flex-col items-center gap-3 sm:flex-row sm:justify-between sm:items-center">
                   <Select
-                    value={selectedFilter || "initial-interview"}
-                    onValueChange={(value) => {
-                      setSelectedFilter(value)
-                      if (value === "resume-screening") {
-                        window.location.href = "/applicants/jobdetails/leaddeveloper/LeadDeveloperRS/"
-                      } else if (value === "phone-call") {
-                        window.location.href = "/applicants/jobdetails/leaddeveloper/LeadDeveloperPI"
-                      } else if (value === "shortlisted") {
-                        window.location.href = "/applicants/jobdetails/leaddeveloper/LeadDeveloperSL"
-                      } else if (value === "assessments") {
-                        window.location.href = "/applicants/jobdetails/leaddeveloper/LeadDeveloperAS"
-                      } else if (value === "final-interview") {
-                        window.location.href = "/applicants/jobdetails/leaddeveloper/LeadDeveloperFI"
-                      } else if (value === "job-offer") {
-                        window.location.href = "/applicants/jobdetails/leaddeveloper/LeadDeveloperFJO"
-                        }
-                    }}
-                  >
+  value={selectedFilter || "initialinterview"}
+  onValueChange={(value) => {
+    setSelectedFilter(value)
+    if (jobtitle) {
+      navigate(`/applicants/job/${jobtitle}/${value}`)
+    }
+  }}
+>
+
                     <SelectTrigger className="w-40 border-none shadow-none font-bold text-black text-sm">
-                      <SelectValue placeholder="Initial Interview" />
+                      <SelectValue placeholder="Shortlisted" />
                     </SelectTrigger>
                     <SelectContent>
-                      <SelectItem className="font-bold" value="resume-screening">
+                      <SelectItem className="font-bold" value="resumescreening">
                         Resume Screening
                       </SelectItem>
-                      <SelectItem value="phone-call" className="font-bold">
+                      <SelectItem value="phonecallinterview" className="font-bold">
                         Phone Call Interview
                       </SelectItem>
                       <SelectItem value="shortlisted" className="font-bold">
                         Shortlisted
                       </SelectItem>
-                      <SelectItem value="initial-interview" className="font-bold">
+                      <SelectItem value="initialinterview" className="font-bold">
                         Initial Interview
                       </SelectItem>
                       <SelectItem value="assessments" className="font-bold">
                         Assessments
                       </SelectItem>
-                      <SelectItem value="final-interview" className="font-bold">
+                      <SelectItem value="finalinterview" className="font-bold">
                         Final Interview
                       </SelectItem>
-                      <SelectItem value="job-offer" className="font-bold">
+                      <SelectItem value="forjoboffer" className="font-bold">
                         For Job Offer
                       </SelectItem>
-                      <SelectItem value="offer-finalization" className="font-bold">
+                      <SelectItem value="offerfinalization" className="font-bold">
                         For Offer and Finalization
                       </SelectItem>
                       <SelectItem value="onboarding" className="font-bold">
